move convertGqltoArg functions to helpers
